// client/src/socket/message-action.ts
export enum MessageAction {
  OK = "ok",
  AUTH_USER = "auth-user",
  CREATE_DOC = "create-doc",
  OPEN_DOC = "open-doc",
  EDIT_DOC = "edit-doc",
  OPEN_DOCS = "open-docs",
  UPDATE_DOC_ACCESS = "update-doc-access",
  UPDATE_DOC_NAME = "update-doc-name",
  ERR_AUTH = "err-auth",
  ERR_AUTH_REQUIRED = "err-auth-required",
  ERR_SOCKET_DOC_OPEN = "err-socket-doc-open",
  ERR_NO_DOC = "err-no-doc",
  ERR_PATCH_APPLY = "err-patch-apply",
  CLOSE_DOC = "close-doc",
  DELETE_DOC = "delete-doc",
  FOCUS_DOC = "focus-doc",
}